chore: move modal subsections into their own package
