+++ src/views/app/detail-panel/preview/parts/info-details-modal/tests/smime-subsection.test.tsx
@@ -8,8 +8,8 @@ import React from 'react';
 
 import { screen } from '@testing-library/react';
 
-import { setupTest } from '../../../../../../carbonio-ui-commons/test/test-setup';
-import { MessageSignature } from '../../../../../../types';
+import { setupTest } from '../../../../../../../carbonio-ui-commons/test/test-setup';
+import { MessageSignature } from '../../../../../../../types';
 import { SmimeSubsection } from '../smime-subsection';
 
 describe('SmimeSubsection', () => {
